backend/requirements.txt
# ============================================
# Web 框架
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ============================================
# 数据库
# ============================================
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# ============================================
# Redis
# ============================================
redis[hiredis]>=5.0.0

# ============================================
# Excel 处理
# ============================================
openpyxl>=3.1.2

# ============================================
# 图片处理
# ============================================
Pillow>=10.2.0
opencv-python-headless>=4.9.0
numpy>=1.26.0

# ============================================
# OCR
# ============================================
rapidocr-onnxruntime>=1.3.0

# ============================================
# HTTP 客户端
# ============================================
httpx>=0.26.0

# ============================================
# 数据验证
# ============================================
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================
# 开发工具
# ============================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.0.0
flake8>=7.0.0
